feat:separate codebase and fix screen color
