+++ main.ts
@@ -1,6 +1,6 @@
 import {httpServer} from './server/sever'
 import WebSocket from "ws";
-import { onConnect } from './websocket';
+import { onConnect } from './ws/websocket';
 
 const HTTP_PORT = 3000;
 
@@ -14,6 +14,5 @@ const wsServer = new WebSocket.Server({ port: 8080 }, () => {
 });
 
 wsServer.on("connection", onConnect);
-
