+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "npx ts-node servers",
+    "start": "npx ts-node main",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
