+++ streams/write.ts
